Update gen002140.sh
The quote marks will cause the "for BADSHELL in '...' " statement to put the entire contents within the quotes into BADSHELL instead of each of the space delimited strings as intended, so nothing will ever match in the following grep statement. This has some bad repercussions later; for example, /dev/null perms will get changed to 755 in gen002220.sh (default is 666) which will break scp and sftp (took me a few hours to figure that out), among other things.